Start glm split
